% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcirt-package.R
\docType{package}
\name{fcirt-package}
\alias{fcirt-package}
\title{The 'fcirt' package.}
\description{
Forced Choice in Item Response Theory
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
