% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extract.R
\name{extract}
\alias{extract}
\title{results extraction}
\usage{
extract(x, pars)
}
\arguments{
\item{x}{returned object}

\item{pars}{Names of extracted parameters. They can be "theta" (person trait estimates), "alpha" (statement discrimination parameters), "delta" (statement location parameters), "tau" (statement threshold parameters), data" (fcirt.Data), "fit" (the stanfit object), and "dimension" (the input column vector mapping each statement to each trait).}
}
\value{
Selected results output
}
\description{
This function extracts estimation results.
}
\examples{
Data <- c(1,1,0,1)
Data <- matrix(Data,nrow = 2)
pairmap <- c(1,3,2,4)
pairmap <- matrix(pairmap,nrow = 2)
ind <- c(1,2,3,4)
ParInits <- c(1, 1, 1, 1, 1, -1, 1, 1, -1, -1, -1, -1)
ParInits <- matrix(ParInits, ncol = 3)
mod <- fcirt(fcirt.Data=Data,pairmap=pairmap,ind=ind,ParInits=ParInits,iter=5,chains=1)
alpha <- extract(mod, 'alpha')
}
